check for dropTarget not being null in addDropTarget
There is a problem with Drag and Drop interaction with echopointng Tree. When a tree node is collapsed and nodes from the branch contain drop targets the targets may not be found in the DOM tree any longer. When the ExtrasDragSource.MessageProcessor
processes init messages in processInit() to add the drop targets a null element is referenced. I changed the code to check that dropTarget isn't null. I hit this back in 2009 on a school project before this project was in github. I recently was working on the project porting to maven and saw these files were in maven central, but hit the problem again.